+++ include/kernel/console.h
@@ -13,11 +13,22 @@ typedef struct console_driver {
 void console_register(const console_driver_t * driver);
 void console_init_all(void);
 
+void console_locking_init(void);
+void console_lock(void);
+void console_unlock(void);
+
 void console_putc(char c);
 void console_write(const char *text);
 void console_writeln(const char *text);
 void console_write_hex32(uint32_t value);
 void console_write_u32_dec(uint32_t value);
 
-#endif
+void console_raw_putc(char c);
+void console_raw_write(const char *text);
+void console_raw_writeln(const char *text);
+void console_raw_write_hex32(uint32_t value);
+void console_raw_write_u32_dec(uint32_t value);
 
+void console_lock_test_once(void);
+
+#endif